requirements.txt
# Core dependencies
langgraph>=0.0.40
langchain>=0.1.0
openai>=1.0.0
requests>=2.31.0
numpy>=1.24.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
nltk>=3.8.0

# File processing
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0

# Additional utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
pandas>=2.0.0

streamlit>=1.28.0
plotly>=5.15.0
pandas>=2.0.0
altair>=5.0.0

# Additional utilities
urllib3>=1.26.0
pathlib2>=2.3.0